move electricit data filter to own models
