+++ src/app/models/electricityFilter.ts
@@ -1,0 +1,6 @@
+export interface ElectricityDataFilter {
+    filter: boolean,
+    display: string,
+    id: number,
+    name: string
+  }
