+++ src/app/models/idb.ts
@@ -1,3 +1,4 @@
+import { ElectricityDataFilter } from './electricityFilter';
 
 
 export interface IdbAccount {
@@ -24,7 +25,10 @@ export interface IdbFacility {
     size: number,
     units: string,
     division: string,
-    img: string
+    img: string,
+    tableElectricityFilter?: ElectricityDataFilter,
+    electricityInputFilter?: ElectricityDataFilter
+    
 }
 
 export interface IdbUtilityMeterGroup {
