+++ src/app/utility/energy-consumption/utility-meter-data/electricity-data-table/electricity-data-table.component.ts
@@ -1,8 +1,8 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { Subscription } from 'rxjs';
-import { UtilityMeterDatadbService } from 'src/app/indexedDB/utilityMeterData-db.service';
+import { ElectricityDataFilter } from 'src/app/models/electricityFilter';
 import { IdbUtilityMeter, IdbUtilityMeterData } from 'src/app/models/idb';
-import { ElectricityDataFilter, UtilityMeterDataService } from '../utility-meter-data.service';
+import { UtilityMeterDataService } from '../utility-meter-data.service';
 
 @Component({
   selector: 'app-electricity-data-table',
@@ -27,7 +27,7 @@ export class ElectricityDataTableComponent implements OnInit {
   electricityDataFilters: Array<ElectricityDataFilter>;
   electricityDataFilterSub: Subscription;
   allChecked: boolean;
-  constructor(private utilityMeterDataService: UtilityMeterDataService, private utilityMeterDataDbService: UtilityMeterDatadbService) { }
+  constructor(private utilityMeterDataService: UtilityMeterDataService) { }
 
   ngOnInit(): void {
     if (this.meterListItem.meterDataItems.length != 0) {
